Revert "refactor: move JsonRenderPlugin into private ns"
This reverts commit 60719aabd31255ff315eb6dda768478dee371b8f.
